refactor: 컨트롤러 리팩터링
