+++ src/main/java/roomescape/controller/MemberController.java
@@ -1,13 +1,14 @@
 package roomescape.controller;
 
 import java.util.List;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import roomescape.service.LoginMemberService;
 import roomescape.service.dto.LoginMemberResponse;
 
-@Controller //rename
+@RestController
+@RequestMapping("/members")
 public class MemberController {
     private final LoginMemberService loginMemberService;
 
@@ -15,13 +16,7 @@ public class MemberController {
         this.loginMemberService = loginMemberService;
     }
 
-    @GetMapping("/reservation") //이동 ^^
-    public String findReservationPage() {
-        return "reservation";
-    }
-
-    @GetMapping("/members")
-    @ResponseBody //restController 로 변경
+    @GetMapping
     public List<LoginMemberResponse> findAllMembers() {
         return loginMemberService.findAll();
     }
