+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,46 +1,14 @@
 package roomescape.controller;
 
-import jakarta.validation.Valid;
-import java.net.URI;
-import java.util.List;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import roomescape.domain.LoginMember;
-import roomescape.service.ReservationService;
-import roomescape.service.dto.ReservationRequest;
-import roomescape.service.dto.ReservationResponse;
 
-@RestController
-@RequestMapping("/reservations")
+@Controller
+@RequestMapping("/reservation")
 public class ReservationController {
-    private final ReservationService reservationService;
-
-    public ReservationController(final ReservationService reservationService) {
-        this.reservationService = reservationService;
-    }
-
     @GetMapping
-    public List<ReservationResponse> findAllReservations() {
-        return reservationService.findAll();
-    }
-
-    @PostMapping
-    public ResponseEntity<ReservationResponse> createReservation(
-            @RequestBody @Valid final ReservationRequest reservationRequest, LoginMember loginMember) {
-        ReservationResponse reservationResponse = reservationService.create(reservationRequest, loginMember);
-        return ResponseEntity.created(URI.create("/reservations/" + reservationResponse.id()))
-                .body(reservationResponse);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteReservation(@PathVariable final long id) {
-        reservationService.deleteById(id);
-        return ResponseEntity.noContent().build();
+    public String findReservationPage() {
+        return "reservation";
     }
 }
